Add `@ts-ignore` to suppress TypeScript error in model template
The added `@ts-ignore` directive resolves a TypeScript error when calling `store.getByField` in the generated model template.
This mostly to achieve a quick fix which should not be the final solution.
The root problem is that Entity and BaseEntity does not allow other types than string for id, so the Compact and CompactEntity are an Omit from Entity with an override of the id field.